requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.4
astroid==2.5.1
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.3
django-allauth==0.44.0
django-crispy-forms==1.11.2
django-decouple==2.1
django-heroku==0.0.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
inflection==0.5.1
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.1
mccabe==0.6.1
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.1.0
pylint==2.7.2
pyparsing==2.4.7
python-dateutil==1.5
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.5
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
svgwrite==1.4.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Tree==0.2.4
uritemplate==3.0.1
urllib3==1.26.4
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
